+++ src/models/StyledComponent.js
@@ -24,6 +24,11 @@ import type { RuleSet, Target } from '../types'
 // an empty execution context every single time...
 const STATIC_EXECUTION_CONTEXT = {}
 
+type BaseState = {
+  theme?: ?Theme,
+  generatedClassName?: string,
+}
+
 export default (ComponentStyle: Function, constructWithOptions: Function) => {
   const identifiers = {}
 
@@ -54,11 +59,12 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
       : componentId
   }
 
-  class BaseStyledComponent extends Component {
+  class BaseStyledComponent extends Component<*, BaseState> {
     static target: Target
     static styledComponentId: string
     static attrs: Object
     static componentStyle: Object
+    static defaultProps: Object
     static warnTooManyClasses: Function
 
     attrs = {}
@@ -176,10 +182,10 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
         return
       }
 
-      this.setState(oldState => {
+      this.setState(prevState => {
         const theme = determineTheme(
           nextProps,
-          oldState.theme,
+          prevState.theme,
           this.constructor.defaultProps
         )
         const generatedClassName = this.generateAndInjectStyles(
@@ -213,7 +219,7 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
         .filter(Boolean)
         .join(' ')
 
-      const baseProps = {
+      const baseProps: any = {
         ...this.attrs,
         className,
       }
@@ -281,7 +287,7 @@ export default (ComponentStyle: Function, constructWithOptions: Function) => {
         ]),
       }
 
-      static withComponent(tag) {
+      static withComponent(tag: Target) {
         const { componentId: previousComponentId, ...optionsToCopy } = options
 
         const newComponentId =