+++ src/models/StyledNativeComponent.js
@@ -12,11 +12,17 @@ import { CHANNEL, CHANNEL_NEXT, CONTEXT_CHANNEL_SHAPE } from './ThemeProvider'
 import type { Theme } from './ThemeProvider'
 import type { RuleSet, Target } from '../types'
 
+type State = {
+  theme?: ?Theme,
+  generatedStyles: any,
+}
+
 export default (constructWithOptions: Function, InlineStyle: Function) => {
-  class BaseStyledNativeComponent extends Component {
+  class BaseStyledNativeComponent extends Component<*, State> {
     static target: Target
     static styledComponentId: string
     static attrs: Object
+    static defaultProps: Object
     static inlineStyle: Object
     root: ?Object
 
@@ -94,10 +100,10 @@ export default (constructWithOptions: Function, InlineStyle: Function) => {
       theme?: Theme,
       [key: string]: any,
     }) {
-      this.setState(oldState => {
+      this.setState(prevState => {
         const theme = determineTheme(
           nextProps,
-          oldState.theme,
+          prevState.theme,
           this.constructor.defaultProps
         )
         const generatedStyles = this.generateAndInjectStyles(theme, nextProps)
@@ -121,7 +127,8 @@ export default (constructWithOptions: Function, InlineStyle: Function) => {
         console.warn(
           'setNativeProps was called on a Styled Component wrapping a stateless functional component. ' +
             'In this case no ref will be stored, and instead an innerRef prop will be passed on.\n' +
-            `Check whether the stateless functional component is passing on innerRef as a ref in ${displayName}.`
+            `Check whether the stateless functional component is passing on innerRef as a ref in ${displayName ||
+              'UnknownStyledNativeComponent'}.`
         )
       }
     }
@@ -158,6 +165,7 @@ export default (constructWithOptions: Function, InlineStyle: Function) => {
         !isStyledComponent(target) &&
         // NOTE: We can't pass a ref to a stateless functional component
         (typeof target !== 'function' ||
+          // $FlowFixMe TODO: flow for prototype
           (target.prototype && 'isReactComponent' in target.prototype))
       ) {
         propsForElement.ref = this.onRef
@@ -193,7 +201,7 @@ export default (constructWithOptions: Function, InlineStyle: Function) => {
         [CHANNEL_NEXT]: CONTEXT_CHANNEL_SHAPE,
       }
 
-      static withComponent(tag) {
+      static withComponent(tag: Target) {
         const { displayName: _, componentId: __, ...optionsToCopy } = options
         const newOptions = {
           ...optionsToCopy,